Fix NN policy/value contract, UCI protocol bugs, and MCTS correctness issues
- Policy tensor size corrected to 4672 (AlphaZero 73-plane encoding) to
  match training/model.py; was hardcoded 4096, misaligning every batch
  item after the first and zeroing priors for moves from the 8th rank.
  Runtime check now requires the exact expected tensor size.
- Convert the model's tanh value output [-1, 1] to the win probability
  [0, 1] that MCTS consumes; a "draw" was previously read as a loss.
- UCI: "position startpos" and "ucinewgame" rebuild a fresh engine
  instead of reset() (which returned to the constructed FEN, not the
  start position); illegal moves in "position ... moves" now warn and
  stop instead of silently desyncing; "go" parses winc/binc and no
  longer derives a budget from a missing clock value.
- MCTS: VirtualLossGuard is move-only (implicit copy double-decremented
  virtual loss); search threads inherit game history so repetition
  detection sees pre-root positions; time/simulation budget checked per
  pipeline slot instead of once per batch.
- NN: session_mutex serializes ONNX session use between the batch
  worker's hot-reload and measure_latency_ms().
- TuningParameters members zero-initialized; an old fenrir.cfg missing
  BatchTimeoutMs read uninitialized memory.
- full_moves widened to uint16_t (overflowed at move 255).
- Self-play loop uses one get_terminal_state() call per position
  instead of three full move generations.
- Tests updated to the refactored Tree/Node/NN API; real-model NN test
  asserts value range instead of an exact stub value.
- README brought up to date with engine features.

+++ include/chess/board.h
@@ -48,7 +48,7 @@ namespace chess
         uint8_t en_passant_square; /* 64 = none */
         uint8_t color;
         uint8_t half_move_clock;
-        uint8_t full_moves;
+        uint16_t full_moves; /* uint8_t overflows in games longer than 255 full moves */
     };
 
     /*
@@ -79,7 +79,7 @@ namespace chess
 
         uint8_t color;
         uint8_t half_move_clock;
-        uint8_t full_moves;
+        uint16_t full_moves;
         Fen fen;
 
         void build_board(const std::string &placement);
